PROTECTED_FIELDS
auth permission added livewire wise too

auth permission added livewire wise too

auth permission added livewire wise too

auth permission added livewire wise too

Update api_v1.php

+++ app/Livewire/Account/Note/Table.php
@@ -36,6 +36,7 @@ class Table extends Component
 
     public function delete($id)
     {
+        abort_unless(auth()->user()?->can('account note.delete'), 403);
         $model = AccountNote::find($id);
         if ($model) {
             $model->delete();
